feat: implement Document Sequencing (Oracle Fusion GL > Setup > Document Sequencing)
Oracle Fusion Financials feature for automatic sequential document numbering
for regulatory compliance. Supports gapless and gap-permitted sequences,
configurable reset frequency (daily/monthly/quarterly/annually), prefix/suffix
formatting, sequence assignments to document categories, and full audit trail.

New files:
- migrations/144_document_sequencing.sql: DB tables (sequences, assignments, audit)
- handlers/document_sequencing.rs: REST API handlers for full CRUD + generation
- tests/suite/document_sequencing.rs: 25 e2e tests covering:
  - Sequence CRUD and lifecycle (create, list, get, activate, deactivate, delete)
  - Sequence assignments (create, list, deactivate, delete)
  - Number generation (via assignment and direct)
  - Audit trail queries
  - Dashboard summary
  - Validation edge cases (empty code, invalid type, duplicate code)
  - Full end-to-end flow

Modified:
- state.rs: Add DocumentSequencingEngine to AppState
- handlers/mod.rs: Register document_sequencing module and API routes
- helpers.rs: Add engine to test state, cleanup for doc sequencing tables
- workflow_helpers.rs: Add engine to workflow test state
- mod.rs: Register test module
- repository.rs: Fix SUM(bigint) cast to avoid numeric type mismatch

+++ crates/atlas-core/src/document_sequencing/repository.rs
@@ -423,7 +423,7 @@ impl DocumentSequencingRepository for PostgresDocumentSequencingRepository {
                 COUNT(*) FILTER (WHERE status = 'active') as active,
                 COUNT(*) FILTER (WHERE sequence_type = 'gapless') as gapless,
                 COUNT(*) FILTER (WHERE sequence_type = 'gap_permitted') as gap_permitted,
-                COALESCE(SUM(current_value), 0) as total_generated
+                COALESCE(SUM(current_value), 0)::bigint as total_generated
             FROM _atlas.document_sequences WHERE organization_id = $1"#,
         )
         .bind(org_id).fetch_one(&self.pool).await
